medical-triage-backend/requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
reportlab
pydantic
anthropic
twilio
python-dotenv
mangum 